Move specific implementations of PaillierParams into CGGMP parameters module
Since the test params must have some values set to accommodate CGGMP requirements,
which is out of scope of `paillier` submodule.
